feat: refactor payment handling and introduce new payment URL generation flow

+++ PsychologySupport/Services/Payment/Payment.Application/Payments/Dtos/VnPayCallbackDto.cs
@@ -20,8 +20,7 @@ public record VnPayCallbackDto
     public string? OrderInfo { get; set; }
 
     [FromQuery(Name = "vnp_Amount")]
-    public float Amount { get; set; }
+    public decimal Amount { get; set; }
     
-
     public bool Success => "00".Equals(ResponseCode);
 }
